Improve test failure modes
Previously, the test mocked the Riemann event builder by specifying the
values that the method should be called with. This resulted in the test
throwing a NullPointerException if you called the Riemann event builder
with a different value than what the test was expecting.

This changes the test to allow any string for the service and state
methods on the Riemann event builder, and verify later that the methods
were called with the expected values.