Include JUnit test results in notifications if available
In the GoCD UI, GoCD automatically generates a "tests" results page for
you if you generate a test artifact with the name `results.xml` that
conforms to the JUnit results spec.

This PR adds support for including the contents of that artifact with
the Riemann notification by looking for said test artifact, performing
an XML -> JSON transform on it, and including the result in the
notification if possible.