Only set basic auth HTTP header on requests to GoCD server
Previously, this was setting the basic auth header with the plugin's
provided username and password for all HTTP requests that the GoCD
server JVM was making. This has the possibility to leak the credentials
externally.

This change makes it so that we explicitly set the basic auth header on
all requests (instead of setting the default Authentication mechanism),
and ensures that it is only set on requests to localhost.